Fix: Refactors command argument separation.
What I had before was much more complicated and wrong. I followed some piece
of Mathworks documentation that I cannot find anymore and it was wrong. It's
not the first time that the documentation is wrong. But now I implemented
according to what I see in MATLAB itself and the rule is actually much
simpler (because it's a weird language that accepts some weird things).
Now it correctly parses `minimize max( D(1,2), D(2,1) )`.