Fix: Strings cannot span multiple lines.
If the file has a string that is not properly quoted, it would span the string
across multiple lines, messing with the highlight of otherwise valid code and
making it very hard to spot the problem, since the error would be generated far
from the actual culprit. This fixes that, so the string gets terminated on a
new line, the highlight does not get messed up and the error is on the right line.