Fix: Don't downcast lexer->lookahead to char, use uint32_t everywhere.
Tree-Sitter supports UTF-8 so `lexer->lookahead` can be multibyte. It was causing
problems when multibyte codepoints that had the last byte with the same number
as `%` or `\0` or `'` (for example) being considered those chars and incorrectly
stoping at wrong points, even causing infite loops because some of them are
detected but never consumed.