fix: don't consume newline when looking for line continuation
Fixes !29

A line continuation makes the next line a continuation of the current
one. However, the newline from the previous line was being skipped by
the external scanner, making the line continuation become a continuation
of the previous line (when the only thing in the current line was the
line continuation). Since the function definition requries a newline at
the end of the `function` line, it was breaking with the new example.

This fix should make the line continuation behave as intended.