fix: identifier after newline
Closes #44

The catch clause seems to be only one that causes problem because of an
optional identifier that can come after it. I had forbidden the
identifier scanner to consume new lines because of this, but this broke
other constructs, and this commit now gives it it's own symbol to allow
the scanner to identify if this is a capture or an identifier.