fix: fixes matrices and cells not allowing new-lines after the opening char
Fixes #44

My first attempt was to modify `identifier` and make it consume
newlines. However, it stops the internal scanner from consuming
new-lines, which broke other code. Another solution is then to
explicitely say that a new-line may come after [ or {, so the internal
scanner consumes them and the external scanner only consumes the
identifier itself. This solution did not break any other test.

A test for class properties was also added, because it was missing and
it was the one that showed the regression.