fix: Do not consume empty lines after ...
In this situation:

```
v = 'a';
switch v
  case 'a' ...

    disp('here');
end
```

The line continuation should not go past the empty line after it.
However, in

```
[ 1 ...

2 ...

]
```

It should. This commit does that. It changes the tree in some tests,
moving the line continuation out of the node they were before, but still
parses correctly.

Fixes #104